% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt.R
\name{lt}
\alias{lt}
\title{Specification of covariates entering the long-term part
in a promotion time cure model.}
\usage{
lt(...)
}
\description{
Specification of covariates entering the long-term part
in a promotion time cure model.
}
\keyword{internal}
