% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_amlps_print.R
\name{print.amlps}
\alias{print.amlps}
\title{Print an additive partial linear model object.}
\usage{
\method{print}{amlps}(x, ...)
}
\arguments{
\item{x}{An object of class \code{amlps}.}

\item{...}{Additional arguments to be passed to print.}
}
\description{
Print an additive partial linear model object.
}
\details{
Prints informative output of a fitted additive partial linear model.
 In particular, the model formula, sample size, number of B-splines in basis,
 number of smooth terms, the chosen penalty order,
 the latent field dimension, the estimated coefficients of the linear part,
 the estimated standard deviation of the error and the effective dimension
 of the smooth terms are displayed.
}
\seealso{
\code{\link{amlps}}, \code{\link{amlps.object}},
 \code{\link{plot.amlps}}
}
\author{
Gressani Oswaldo \email{oswaldo_gressani@hotmail.fr}.
}
