% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_coxlps_print.R
\name{print.coxlps}
\alias{print.coxlps}
\title{Print a coxlps object.}
\usage{
\method{print}{coxlps}(x, ...)
}
\arguments{
\item{x}{An object of class \code{coxlps}.}

\item{...}{Further arguments passed to print.}
}
\description{
Print method for a \code{coxlps} object.
}
\details{
Prints informative output of a fitted Cox proportional hazards model
 with Laplace-P-splines.
}
\seealso{
\code{\link{coxlps}}
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}.
}
