% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_gamlps_print.R
\name{print.gamlps}
\alias{print.gamlps}
\title{Print a generalized additive model object.}
\usage{
\method{print}{gamlps}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gamlps}.}

\item{...}{Further arguments to be passed to print.}
}
\description{
Print a generalized additive model object.
}
\details{
Prints informative output of a fitted generalized additive model.
 In particular, the model formula, sample size, number of B-splines in basis,
 number of smooth terms, the chosen penalty order,
 the latent field dimension, model degrees of freedom, the estimated
 coefficients of the linear part and
 the estimated effective degrees of freedom of the smooth terms are
 displayed.
}
\seealso{
\code{\link{gamlps}}, \code{\link{gamlps.object}},
 \code{\link{plot.gamlps}}
}
\author{
Gressani Oswaldo \email{oswaldo_gressani@hotmail.fr}.
}
