% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st.R
\name{st}
\alias{st}
\title{Specification of covariates entering the short-term part
in a promotion time cure model.}
\usage{
st(...)
}
\description{
Specification of covariates entering the short-term part
in a promotion time cure model.
}
\keyword{internal}
